refactor: address PR #1 review comments
- Rename package app.teambalance → com.github.zzave.teambalance.api
- Extract dependency versions (wirespec, testcontainers, archunit) to gradle.properties
- Upgrade Redis from 7-alpine to 8-alpine in docker-compose.yml
- Replace Makefile help target with enhanced version supporting above-line descriptions
