+++ api/src/main/kotlin/com/github/zzave/teambalance/api/interfaces/HealthController.kt
@@ -1,4 +1,4 @@
-package app.teambalance.interfaces
+package com.github.zzave.teambalance.api.interfaces
 
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RestController
