+++ gradle.properties
@@ -2,3 +2,8 @@ kotlin.code.style=official
 org.gradle.parallel=true
 org.gradle.caching=true
 org.gradle.jvmargs=-Xmx2g -XX:+UseParallelGC
+
+# Dependency versions
+wirespecVersion=0.14.3
+testcontainersVersion=1.20.4
+archunitVersion=1.3.0
